WIP: develop 병합 전 임시 커밋

+++ src/main/java/com/example/delivery/domain/menu/service/MenuServiceImpl.java
@@ -1,5 +1,7 @@
 package com.example.delivery.domain.menu.service;
 
+import com.example.delivery.common.exception.CustomException;
+import com.example.delivery.common.exception.enums.ErrorCode;
 import com.example.delivery.domain.store.entity.Store;
 import com.example.delivery.domain.store.repository.StoreRepository;
 import com.example.delivery.domain.menu.dto.RequestDto;
@@ -19,14 +21,23 @@ public class MenuServiceImpl implements MenuService {
     private final MenuRepository menuRepository;
     private final StoreRepository storeRepository;
 
+    public Menu findMenuOrThrow(Long menuId) {
+        return menuRepository.findById(menuId)
+                .orElseThrow(() -> new CustomException(ErrorCode.MENU_NOT_FOUND));
+    }
+    
     @Override
     @Transactional
-    public ResponseDto create(Long loginUserId, RequestDto requestDto) {
+    public ResponseDto create(Long ownerId, RequestDto requestDto) {
 
-//        Store store = storeRepository.findById(requestDto.getId())
-//                .orElseThrow(() -> new NotFoundException(ErrorCode.STORE_NOT_FOUND));
-        Store store = storeRepository.findMyStoreOrElseThrow(requestDto.getId(), loginUserId);
-        // 메뉴 생성은 사장님만 가능하게 해야되는데 사장님 검증 로직이 빠짐
+        // 1. 가게 조회
+        Store store = storeRepository.findByIdOrElseThrow(requestDto.getId());
+
+        // 2. 가게 오너 검증
+        store.validateOwner(ownerId);
+
+        //Store store = storeRepository.findMyStoreOrElseThrow(requestDto.getId(), loginUserId);
+        // 메뉴 생성은 사장님만 가능하게 해야되는데 사장님 검증 로직이 빠짐 = > 문제점 : repository가 조회 + 검증을 함, 즉 비즈니스 로직을 처리함
 
         Menu menu = Menu.of(store, requestDto.getName(), requestDto.getPrice(), requestDto.getDescription());
         menuRepository.save(menu);
@@ -36,12 +47,16 @@ public class MenuServiceImpl implements MenuService {
 
     @Override
     @Transactional
-    public ResponseDto update(Long loginUserId, Long menuId, RequestDto requestDto) {
+    public ResponseDto update(Long ownerId, Long menuId, RequestDto requestDto) {
         // 사장님인지 검증 로직
         // 메뉴가 속한 가게의 소유자가 로그인된 유저인걸 검증하는 메서드
-        storeRepository.findMyStoreOrElseThrow(requestDto.getId(), loginUserId);
-        //메뉴 조회
-        Menu menu = menuRepository.findByIdOrElseThrow(loginUserId);
+       // storeRepository.findMyStoreOrElseThrow(requestDto.getId(),ownerId);
+        // 1. 메뉴 조회
+        Menu menu = menuRepository.findByIdOrElseThrow(ownerId);
+        // 2. 스토어 오너 검증
+        Store store = menu.getStore();
+        store.validateOwner(ownerId);
+
         // 메뉴 업데이트
         menu.update(requestDto.getName(), requestDto.getPrice(), requestDto.getDescription());
 
@@ -50,11 +65,13 @@ public class MenuServiceImpl implements MenuService {
 
     @Override
     @Transactional
-    public void delete(Long loginUserId, Long menuId) {
+    public void delete(Long ownerId, Long menuId) {
         // 메뉴 조회
-        Menu menu = menuRepository.findByIdOrElseThrow(loginUserId);
+        Menu menu = menuRepository.findByIdOrElseThrow(ownerId);
         // 사장인지 검증
-        storeRepository.findMyStoreOrElseThrow(menu.getStore().getId(), loginUserId);
+        // storeRepository.findMyStoreOrElseThrow(menu.getStore().getId(), ownerId);
+        Store store = menu.getStore();
+        store.validateOwner(ownerId);
         menu.delete();
     }
     @Override
